pythonBackend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.5=py310h9128911_0
bzip2=1.0.8=he774522_0
ca-certificates=2022.10.11=haa95532_0
cachetools=5.2.0=pypi_0
certifi=2022.9.24=py310haa95532_0
charset-normalizer=2.1.1=pypi_0
click=8.0.4=py310haa95532_0
colorama=0.4.5=py310haa95532_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flask=2.0.3=pyhd3eb1b0_0
flask-cors=3.0.10=pypi_0
flask-sock=0.5.2=pypi_0
google-api-core=2.10.2=pypi_0
google-api-python-client=2.66.0=pypi_0
google-auth=2.14.1=pypi_0
google-auth-httplib2=0.1.0=pypi_0
googleapis-common-protos=1.57.0=pypi_0
h11=0.14.0=pypi_0
httplib2=0.21.0=pypi_0
idna=3.4=pypi_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.2=py310haa95532_0
ipython=8.4.0=py310haa95532_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.18.1=py310haa95532_1
jinja2=3.1.2=py310haa95532_0
joblib=1.2.0=pypi_0
jupyter_client=7.3.5=py310haa95532_0
jupyter_core=4.11.1=py310haa95532_0
libffi=3.4.2=hd77b12b_4
libsodium=1.0.18=h62dcd97_0
markupsafe=2.1.1=py310h2bbff1b_0
matplotlib-inline=0.1.6=py310haa95532_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
nest-asyncio=1.5.5=py310haa95532_0
nltk=3.7=pypi_0
numexpr=2.8.3=py310hb57aa6b_0
numpy=1.23.4=pypi_0
numpy-base=1.23.3=py310h04254f7_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py310hd77b12b_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=pypi_0
pip=22.2.2=py310haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=4.21.9=pypi_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.9=py310haa95532_0
python=3.10.6=hbb2ffb3_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.1=py310haa95532_0
pywin32=302=py310h2bbff1b_2
pyzmq=23.2.0=py310hd77b12b_0
regex=2022.10.31=pypi_0
requests=2.28.1=pypi_0
rsa=4.9=pypi_0
setuptools=65.4.0=py310haa95532_0
simple-websocket=0.8.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
textblob=0.17.1=pypi_0
tk=8.6.12=h2bbff1b_0
tornado=6.2=py310h2bbff1b_0
tqdm=4.64.1=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
tzdata=2022e=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.13=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py310haa95532_2
wsproto=1.2.0=pypi_0
xz=5.2.6=h8cc25b3_0
zeromq=4.3.4=hd77b12b_0
zlib=1.2.13=h8cc25b3_0
